savepoint

+++ v3/messages.go
@@ -4,14 +4,14 @@ import "sync"
 
 type Messages[M any] struct {
 	sync.RWMutex
-	offset   int64
+	offset   int
 	messages []M
 	used     []int
 }
 
-func (m *Messages[M]) LockForAdd() int64 {
+func (m *Messages[M]) LockForAdd() int {
 	m.Lock()
-	return m.offset + int64(len(m.messages))
+	return m.offset + len(m.messages)
 }
 
 func (m *Messages[M]) AddAndUnlock(msg M, used int) {
@@ -20,19 +20,19 @@ func (m *Messages[M]) AddAndUnlock(msg M, used int) {
 	m.Unlock()
 }
 
-func (m *Messages[M]) Used(id int64, delta int) {
+func (m *Messages[M]) Used(id int, delta int) {
 	m.Lock()
 	m.used[id-m.offset] += delta
 	m.Unlock()
 }
 
-func (m *Messages[M]) Get(id int64) M {
+func (m *Messages[M]) Get(id int) M {
 	m.RLock()
 	defer m.RUnlock()
 	return m.messages[id-m.offset]
 }
 
-func (m *Messages[M]) GetDrop() (drop, dropOffset int64) {
+func (m *Messages[M]) GetDrop() (drop, dropOffset int) {
 	m.RLock()
 	defer m.RUnlock()
 	if len(m.messages) == 0 {
@@ -47,7 +47,7 @@ func (m *Messages[M]) GetDrop() (drop, dropOffset int64) {
 	return drop, m.offset + drop
 }
 
-func (m *Messages[M]) Drop(drop int64) {
+func (m *Messages[M]) Drop(drop int) {
 	m.Lock()
 	n := copy(m.messages, m.messages[drop:])
 	copy(m.used, m.used[drop:])
